+++ OnlineSchoolForKids/Domain/Entities/Content/Enrollment.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities.Content
             public string UserId { get; set; } = string.Empty;
             public string CourseId { get; set; } = string.Empty;
             public DateTime EnrollmentDate { get; set; } = DateTime.UtcNow;
-            public decimal Progress { get; set; } = 0; // 0-100
+            public double Progress { get; set; } = 0; // 0-100
             public bool IsCompleted { get; set; } = false;
             public DateTime? CompletedAt { get; set; }
             public string? LastAccessedLessonId { get; set; }
